+++ playwright/helpers/run-helpers.ts
@@ -6,12 +6,26 @@ export const writeStorageState = async (page): void => {
   await page.context().storageState({ path: dataPath })
 }
 
+export const sleep = (duration: number): Promise<void> => {
+    return new Promise((resolve) => {
+        setTimeout(resolve, duration)
+    })
+}
+
 export const openPlayground = async (page): void => {
   await page.goto(testUrl)
   // close intro animation
   await page.getByRole('button').click()
 }
 
-export const visitAboutScreen = async (page): void => await page.getByTitle('Description').click()
+export const visitAboutScreen = async (page): void => {
+  await page.getByTitle('Description').click()
+  // when click has fully been processed, time can be measured
+  await sleep(10)
+}
 
-export const visitPlayground = async (page): void => await page.getByTitle('Return to Game').click()
+export const visitPlayground = async (page): void => {
+  await page.getByTitle('Return to Game').click()
+  // when click has fully been processed, time can be measured
+  await sleep(10)
+}